// src/types/user.ts
export interface User {
    id: string;
    uuid: string;
    gender: string;
    name: {
        title: string;
        first: string;
        last: string;
    };
    email: string;
    picture: {
        large: string;
        medium: string;
        thumbnail: string;
    };
    isFavorite: boolean;
}